LayoutTests/imported/w3c/web-platform-tests/html/syntax/parsing-html-fragments/the-input-byte-stream-015-expected.txt

 

FAIL A page with no encoding information in HTTP, BOM, XML declaration or meta element will be treated as UTF-8. assert_equals: expected 100 but got 50
